requirements.txt
tensorflow
flask
pillow
numpy
matplotlib
gunicorn
requests
